Bugfix for back buttons on the copyright pages
Fixed by changing the path of the pages to have the language code first.

If the opphavsrett page was loaded without first loading one of the other pages the back buttons would fail.

When App.vue is loaded it generates a few global variables, one of the being langCode. This variable is generated using $route.fullPath.split("/")[1]. It therefore expects the language code to be the second element in fullPath.split array (Since the first character is "/", the first element of the split array becomes empty).

However, the path for the two copyright pages had the language code at the end. When loading the page application from one of those the langCode would therefore be "Opphavsrett" instead of "nb" or "nn".