Selecting an era in the menu now takes you to the era on the timeline, not the article
Selecting an era in the menu resulted in being transported directly to the article for the era. This is not the expected behaviour.

Now selecting a menu item scrolls the timeline to the selected era and selects the button. To navigate to the article you have to press the mouse or press the enter key.